refactor(LottoController) : 실제 로또의 세부 기능을 담당하는 LottoController로 로직 분리
